NEXTFLEX Combo-Seq

Index

combo_seq.Config

combo_seq.MissingReferenceError

combo_seq.align_with_salmon

combo_seq.calculate_read_length_distribution

combo_seq.calculate_salmon_metrics

combo_seq.capture_target_files

combo_seq.create_reference_file

combo_seq.create_target_organism_fasta_file

combo_seq.find_common_rnas

combo_seq.generate_mirna_counts

combo_seq.get_genus_names

combo_seq.get_read_q_score!

combo_seq.make_bowtie2_reference

combo_seq.make_metrics_violin_plot

combo_seq.make_salmon_reference

combo_seq.mirna_discovery_calculation

combo_seq.parse_fastq_files

combo_seq.plot_clustering

combo_seq.plot_fragment_lengths

combo_seq.plot_metrics

combo_seq.plot_mirna_counts

combo_seq.progress_bar_update

combo_seq.remove_files

combo_seq.remove_intermediate_files

combo_seq.trim_adapters

combo_seq.write_common_rna_file

Functions

Documentation for combo_seq.jl

combo_seq.MissingReferenceError — Type

struct MissingReferenceError
    message::String
end

A structure that represents a custom exception for a missing reference.

Missing docstring.

Missing docstring for IncorrectOrganismError. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MissingFastqsError. Check Documenter's build log for details.

combo_seq.Config — Type

struct Config
    need_reference::Bool
    mrna::Union{String, Nothing}
    fasta::Union{String, Nothing}
    transcript::Union{String, Nothing}
    genome::Union{String, Nothing}
    organism::Union{String, Nothing}
    threads::Int
end

A structure that stores the command-line arguments.

Functionality

Contains a check to ensure that a Salmon and miRNA fasta file are passed as arguments.

If they were not, an error will be returned. This is to reduce the chances of an alignment errors downstream.

combo_seq.get_genus_names — Function

Create a file with a list of unique species names given a miRNA mature or hairpin fasta.

combo_seq.create_target_organism_fasta_file — Function

Function takes in mirBase mature miRNA fasta file with data from all available organisms and pulls out only the miRNA data pertaining to the target organism.

combo_seq.capture_target_files — Function

capture_target_files(files_to_capture::AbstractString, directory::AbstractString=".")

List all files in a directory.

Check to see if each file contains the target file's string.

Example

julia> capture_target_files(".txt")
3-element Vector{String}:
 "file1.txt"
 "file2.txt"
 "file3.txt"

combo_seq.progress_bar_update — Function

Update progress bar on the command line.

combo_seq.get_read_q_score! — Function

get_read_q_score!(line::String, q_score_list::Vector{Number})

Calculate the quality score for a given quality string.

This function accepts ASCII-encoded quality score and produces the average q score using a scalar value that estimates the 'total error probability' for that record. It means that if you want to calculate average quality score, you don't just sum up all the phred scores and find the average, but rather consider the probability distribution of the scores.

The Phred score for a base Q is calculated as Q = ASCII value of quality score - 33. The error probability P for that base is then calculated as P = 10^(-Q/10). Then these probabilities are summed up for all the bases to derive total error.

Arguments

q_score - The quality scores of the current line. q_score_list - The vector containing all calculated quality scores.

Returns

The updated q score list the line's average quality score.

Reference

Illumina's explanation on quality scores and their corresponding error rates: https://www.illumina.com/science/technology/next-generation-sequencing/plan-experiments/quality-scores.html

Example

julia> get_read_q_score!("FGFEEE<FC<FGGGGGGGGGGGFFGFG8<@8CFFGF8EFGGCGFGGGGGGFG", [36.2, 35.9])
3-element Vector{Float64}:
 36.2
 35.9
 35.7

combo_seq.make_bowtie2_reference — Function

Make miRNA bowtie2 reference to align fastqs with.

combo_seq.make_salmon_reference — Function

Make Salmon reference for mRNA alignment.

combo_seq.create_reference_file — Function

Function asks the user if they would like to download a non-human reference for use by Salmon during the alignment step. If yes, the files are downloaded and the genome names are extracted to name the Salmon directory.

combo_seq.parse_fastq_files — Function

parse_fastq_files(fastqs::Vector{String}, sample_names::Vector{SubString{String}})

Calculate read counts, percent dimers, and average Q scores for a given set of fastq files.

Parameters

fastqs: A vector of filenames for the fastq files.

sample_names: A vector of sample names corresponding to the fastq files.

Returns

Three dictionaries containing read counts, percent dimers, and average Q scores indexed by sample name.

Example

julia> parse_fastq_files(["sample1.fastq.gz", "sample2.fastq.gz", "sample3.fastq.gz"], 
                        ["sample1", "sample2", "sample3"])
# Returns:
# - A dictionary of read counts.
# - A dictionary of percent dimers.
# - A dictionary of average Q scores.

combo_seq.trim_adapters — Function

trim_adapters(fastqs::Vector{String}, sample_names::Vector{SubString{String}})

Trim the 3' adapter from each read in the provided fastq files.

Combo-Seq Adapters:

5' Adapter: GATCGTCGGACTGTAGAACTCTGAACNNNN

3' Adapter: AAAAAAAAAA

3' bases with a quality score below 20 are trimmed. Reads shorter than 16 bases or those that weren't trimmed are discarded.

Parameters

fastqs: A vector of filenames for the fastq files.

sample_names: A vector of sample names corresponding to the fastq files.

Returns

A list of filenames for the trimmed fastq files.

Example

julia> trim_adapters(["sample1.fastq.gz", "sample2.fastq.gz", "sample3.fastq.gz"], 
                    ["sample1", "sample2", "sample3"])
# Returns:
["sample1.cut.fastq", "sample2.cut.fastq", "sample3.cut.fastq"]

combo_seq.align_with_salmon — Function

align_with_salmon(trimmed_fastq_files::Vector{String}
                    , salmon_reference::AbstractString
                    , sample_names::Vector{SubString{String}}
                    , need_reference::Bool
                    )

This function aligns trimmed fASTQ files to a reference genome using Salmon.

The library type (libType) is set to stranded forward (SF) since the Combo-Seq protocol is stranded in that direction.

Arguments

trimmed_fastq_files: A vector of strings containing the paths to the trimmed fASTQ files.

salmon_reference: A string containing the path to the Salmon reference genome.

sample_names: A vector of strings containing the names of the samples.

need_reference: A string indicating whether or not a custom Salmon index was created.

Returns

The function returns a vector of strings containing the paths to the Salmon output files.

combo_seq.calculate_salmon_metrics — Function

calculate_salmon_metrics(trimmed_fastq_files::Vector{String}
                        , read_count_dict::Dict{String, Int64}
                        , sample_names::Vector{SubString{String}}
                        , dimer_count_dict::Dict{String, Float64}
                        , q_score_dict::Dict{String, Number}
                        )

Align fastq with specified reference RNA types.

Divide reads aligned with total reads to calculate percent alignment. Store alignment information in a dictionary.

#Example

julia> calculate_salmon_metrics("sample1.cut.fastq"
                        , Dict("sample1" => 6390314, "sample2" => 5000000, "sample3" => 7052928))
# Returns:
# - "ComboSeq_metrics.csv"

combo_seq.generate_mirna_counts — Function

generate_mirna_counts(input_sam_file::AbstractString)

This function creates a DataFrame containing all aligned miRNA and the number of times they appear.

Returns

A DataFrame with the following columns:

name: The name of the miRNA.

count: The number of times the miRNA appears in the SAM file.

combo_seq.make_metrics_violin_plot — Function

Create violin plot of each sample's RNA aligment metrics

combo_seq.mirna_discovery_calculation — Function

mirna_discovery_calculation(trimmed_fastq_files::Vector{String}
                            , sample_names::Vector{SubString{String}}
                            , bowtie2_reference_name::AbstractString
                            , organism_name::AbstractString
                            , need_reference::AbstractString
                            )

Align trimmed fastq files to the single organism miRNA bowtie2 reference.

The bowtie2 output SAM is input into function to generate miRNA read counts.

#Example

julia> mirna_discovery_calculation(["sample1.cut.fastq","sample2.cut.fastq","sample3.cut.fastq"]
                                , ["sample1", "sample2", "sample3"])
3-element Vector{DataFrame}:
469x2 DataFrame
 Row │ name             count
     │ String           Int64
─────┼────────────────────────
   1 │ hsa-miR-1307-3p      1
   2 │ hsa-miR-425-5p      58
   3 │ hsa-miR-488-5p       2
 [...]

 3-element Vector{String}:
 "sample1.miRNA.sam"
 "sample2.miRNA.sam"
 "sample3.miRNA.sam"

combo_seq.plot_mirna_counts — Function

plot_mirna_counts(mirna_counts_dfs::Vector{DataFrame}
                    , sample_names::Vector{SubString{String}}
                    )

Plot counts of the number of unique miRNA each sample aligned to.

#Example

julia> plot_mirna_counts([sample1_counts_df,sample2_counts_df,sample3_counts_df]
                        , ["sample1", "sample2", "sample3"])
3-element Vector{String}:
 "sample1_miRNA_counts.csv"
 "sample2_miRNA_counts.csv"
 "sample3_miRNA_counts.csv"

combo_seq.calculate_read_length_distribution — Function

Use aligned SAM files to create a read length distribution file.

combo_seq.plot_fragment_lengths — Function

Create barplot of each sample's fragment lengths based on the number of reads found at each length

combo_seq.plot_metrics — Function

Create barplot of each sample's RNA aligment metrics.

combo_seq.find_common_rnas — Function

find_common_rnas(counts_files::Vector{String}
                , sample_names::Vector{SubString{String}} 
                ; rna_type::String="miRNA", read_count_dict::Dict{String
                , Int64}=Dict{String, Int64}()
                )

Return the RNAs present in all samples, together with their counts and reads per million reads (RPM) or transcripts per million (TPM).

Arguments

counts_files: A vector of files containing sample RNA counts.

sample_names: A vector of sample names.

rna_type: A string specifying the type of RNA ("miRNA" or "mRNA"). Default is "miRNA".

read_count_dict: (Optional for "mRNA" type) A dictionary where keys are sample names and

values are the total number of reads for the corresponding sample. This is required for "miRNA" type to compute RPM values.

Returns

A Tuple of three items:

A list of RNAs that are common to all samples.

A list of dictionaries where each dictionary contains the RNA counts for one sample.

A list of dictionaries where each dictionary contains the RPM (for miRNA) or TPM

(for mRNA) values for the RNAs in one sample.

Example

For miRNA:

julia> find_common_rnas(["sample1_miRNA_counts.csv", "sample2_miRNA_counts.csv"],
                        ["sample1", "sample2"],
                        rna_type="miRNA",
                        read_count_dict=Dict("sample1" => 6390314, "sample2" => 5000000))

combo_seq.write_common_rna_file — Function

write_common_rna_file(rna_names::Base.KeySet{String, Dict{String, Int64}},
                      rna_info::Vector{Dict{String, Number}},
                      normalization_info::Vector{Dict{String, Number}},
                      sample_names::Vector{SubString{String}};
                      rna_type::String="miRNA")

Write the common RNAs found in all samples to an output file. This function generates two sets of files for each sample: one containing raw RNA counts and another containing normalized values (RPM for miRNA or TPM for mRNA). It then combines these files for all samples to produce comprehensive outputs, and subsequently conducts a principal component analysis and UMAP on the common RNA data.

Arguments

rna_names: A set containing the names of RNAs that are common across all samples.

rna_info: A vector of dictionaries with each dictionary containing the RNA counts for one sample.

normalization_info: A vector of dictionaries, each containing the RPM (for miRNA) or TPM (for mRNA) values for the RNAs in one sample.

sample_names: A vector of sample names.

rna_type: (Optional) A string specifying the type of RNA ("miRNA" or "mRNA"). Default is "miRNA".

Outputs

Generates files for each sample named like sample_common_miRNAs.tsv and sample_common_miRNAs_RPM.tsv (or their mRNA equivalents).

Combines these files to produce comprehensive outputs named Common_miRNAs.tsv and Common_RPM_miRNAs.tsv (or their mRNA equivalents).

Conducts principal component analysis and UMAP based on common RNA data.

combo_seq.plot_clustering — Function

plot_clustering(Common_miRNA_File::String)

Generates Principal Component Analysis (PCA) and Uniform Manifold Approximation and Projection (UMAP) plots for libraries given a common RNA file.

This function performs the following steps:

Reads the common RNA file and creates a DataFrame.

Extracts sample names and [mi/m]RNA names.

Ensures that there are at least two samples and more than one miRNA present in the dataset.

Transforms the data using PCA and prepares it for plotting.

If there are at least two principal components, plots the PCA and saves it as "common[mi/m]RNAPCA.png".

If there are at least three principal components, performs UMAP dimensionality reduction and clustering using K-medoids.

Plots the UMAP and saves it as "common[mi/m]RNAUMAP.png".

Writes the PCA and UMAP information to separate CSV files for cluster tracking.

Arguments

Common_[mi/m]RNA_File::String: Path to the common RNA RPM/TPM file.

Outputs

Creates and saves PCA and UMAP plots as "common[mi/m]RNAPCA.png" and "common[mi/m]RNAUMAP.png".

Writes PCA and UMAP information to "PCAinformation.csv" and "UMAPinformation.csv".

combo_seq.remove_files — Function

Spot remove unecessary intermediate files.

combo_seq.remove_intermediate_files — Function

Remove all intermediate files.